feat: declare a canonical URL on every route (MON-269)
Only /quiz emitted `<link rel="canonical">`. Every other page - the
homepage, /deputes, /votes, both detail routes, /groupes, /themes,
/departements, /chat, the share snapshots and the legal pages - shipped
without one, so the same document was independently crawlable on
mon-elu.vercel.app, on every per-deployment *.vercel.app host and on any
future custom domain, with none of them claiming to be the original.
Query-string entries (/quiz?deputy=, /quiz?compare=, /chat?q=) forked
into distinct URLs for the same page on top of that.

- `canonicalUrl()` in `src/lib/site.ts` builds the absolute URL from
  SITE_URL, matching the sitemap's spelling exactly (no trailing slash,
  bare origin for the homepage).
- Every `page.tsx` now declares `alternates: { canonical: ... }`.
  Dynamic routes canonicalise on the resolved identifier, not the URL
  spelling: `themeSlug(data.name)`, the group's canonical slug and
  `data.code` for a department.
- /chat, /mon-depute and /deputes/comparer were client components and
  could not export metadata at all; their bodies move to ChatClient /
  MonDeputeClient (comparer already had a client file), leaving a thin
  server page that carries the metadata.
- `__tests__/app/canonical.test.ts` walks src/app and fails on any
  page.tsx that ships without a canonical or builds one from a hardcoded
  origin. `~offline` and `embed/votes/[id]` are the two allowlisted
  exceptions and are asserted to still exist.
- The Playwright smoke tier asserts the link is actually rendered, on
  desktop-light only since it is viewport-independent.
