+++ frontend/src/app/chat/s/[id]/page.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { api } from '@/lib/api'
 import { ChatAnswerCard } from '@/components/ChatAnswerCard'
-import { SITE_URL } from '@/lib/site'
+import { SITE_URL, canonicalUrl } from '@/lib/site'
 
 // Chat shares are immutable snapshots (ADR-024, mirrors ADR-022 for
 // verifications): this page reads the stored answer via GET /search/share/{id}
@@ -24,6 +24,7 @@ export async function generateMetadata({
   return {
     title,
     description,
+    alternates: { canonical: canonicalUrl(`/chat/s/${id}`) },
     openGraph: { title, description, url: `${SITE_URL}/chat/s/${id}` },
     twitter: { card: 'summary_large_image', title, description },
   }
