+++ frontend/src/app/verifier/v/[id]/page.tsx
@@ -3,7 +3,7 @@ import Link from 'next/link'
 import { notFound } from 'next/navigation'
 import { api } from '@/lib/api'
 import { VerdictCard } from '@/components/VerdictCard'
-import { SITE_URL } from '@/lib/site'
+import { SITE_URL, canonicalUrl } from '@/lib/site'
 
 // Verdicts are immutable snapshots (ADR-022): this page reads the stored
 // verdict via GET /verify/{id} — it must never trigger a new verification.
@@ -32,6 +32,7 @@ export async function generateMetadata({
   return {
     title,
     description,
+    alternates: { canonical: canonicalUrl(`/verifier/v/${id}`) },
     openGraph: { title, description, url: `${SITE_URL}/verifier/v/${id}` },
     twitter: { card: 'summary_large_image', title, description },
   }
