+++ frontend/src/lib/site.ts
@@ -15,3 +15,24 @@ export const SITE_URL = process.env.NEXT_PUBLIC_SITE_URL ?? 'https://mon-elu.ver
 
 /** Host only, for display in OG card footers (e.g. `mon-elu.vercel.app`). */
 export const SITE_HOST = SITE_URL.replace(/^https?:\/\//, '').replace(/\/$/, '')
+
+/**
+ * Absolute canonical URL for a page path (MON-269).
+ *
+ * Every route must declare `alternates: { canonical: canonicalUrl(...) }`.
+ * Without it, the same document is served - and independently crawlable - on
+ * `mon-elu.vercel.app`, on every per-deployment `*-<hash>.vercel.app`, and on
+ * any future custom domain, with none of them claiming to be the original;
+ * query-string variants (`/quiz?deputy=`, `/quiz?compare=`, `/chat?q=`) fork
+ * into distinct URLs for the same page on top of that.
+ *
+ * Absolute rather than relative-plus-`metadataBase` so the emitted URL is
+ * readable as-is in the HTML source. The string must match the corresponding
+ * `sitemap.ts` entry exactly, hence no trailing slash - including on the
+ * homepage, which the sitemap lists as bare `SITE_URL`.
+ */
+export function canonicalUrl(path = '/'): string {
+  const trimmed = path.replace(/\/+$/, '')
+  if (!trimmed) return SITE_URL
+  return `${SITE_URL}${trimmed.startsWith('/') ? trimmed : `/${trimmed}`}`
+}
